// Front-End/src/shared/enums/NotificationType.ts
export enum NotificationType {
  //   BOOKING_REQUEST = "BOOKING_REQUEST",
  //   BOOKING_ACCEPTED = "BOOKING_ACCEPTED",
  //   BOOKING_REJECTED = "BOOKING_REJECTED",
  //   PAYMENT_SUCCESS = "PAYMENT_SUCCESS",
  //   PAYMENT_FAILED = "PAYMENT_FAILED",

  BOOKING_CONFIRMED = "BOOKING_CONFIRMED",
  BOOKING_CANCELLED = "BOOKING_CANCELLED",
  KYC_REQUEST = "KYC_REQUEST",

}